[gh-1552] refactor encoding and decoding, move to moveos_stdlib. (#1560)
* [gh-1552] refactor encoding and decoding, move to moveos_stdlib.

* [gh-1552] add error code to move files.

---------
