+++ Cargo.toml
@@ -146,7 +146,7 @@ enum_dispatch = "^0.3"
 ethereum-types = "0.14.1"
 ethers = { version = "2.0.7", features = ["legacy"] }
 eyre = "0.6.8"
-fastcrypto = { git = "https://github.com/rooch-network/fastcrypto", rev = "643831ec3b67bdd2b5f998c0bec1b7c91823351f" }
+fastcrypto = { git = "https://github.com/rooch-network/fastcrypto", rev = "aa5f9f308b6598779820db8b673050c10cfcc3c1" }
 futures = "0.3.28"
 hex = "0.4.3"
 rustc-hex = "1.0"
