+++ frameworks/moveos-stdlib/doc/README.md
@@ -16,7 +16,9 @@ This is the reference documentation of the MoveOS standard library.
 -  [`0x2::address`](address.md#0x2_address)
 -  [`0x2::any`](any.md#0x2_any)
 -  [`0x2::bag`](bag.md#0x2_bag)
+-  [`0x2::base58`](base58.md#0x2_base58)
 -  [`0x2::bcs`](bcs.md#0x2_bcs)
+-  [`0x2::bech32`](bech32.md#0x2_bech32)
 -  [`0x2::big_vector`](big_vector.md#0x2_big_vector)
 -  [`0x2::copyable_any`](copyable_any.md#0x2_copyable_any)
 -  [`0x2::core_addresses`](core_addresses.md#0x2_core_addresses)
@@ -24,6 +26,7 @@ This is the reference documentation of the MoveOS standard library.
 -  [`0x2::event`](event.md#0x2_event)
 -  [`0x2::features`](features.md#0x2_features)
 -  [`0x2::genesis`](genesis.md#0x2_genesis)
+-  [`0x2::hash`](hash.md#0x2_hash)
 -  [`0x2::hex`](hex.md#0x2_hex)
 -  [`0x2::json`](json.md#0x2_json)
 -  [`0x2::move_module`](move_module.md#0x2_move_module)
