+++ frameworks/rooch-framework/Cargo.toml
@@ -26,8 +26,6 @@ serde_bytes = { workspace = true }
 sha3 = { workspace = true }
 smallvec = { workspace = true }
 hex = { workspace = true }
-bech32 = { workspace = true }
-bs58 = { workspace = true, features = ["check"] }
 http = { workspace = true }
 tracing = { workspace = true }
