+++ frameworks/rooch-framework/doc/README.md
@@ -26,17 +26,14 @@ This is the reference documentation of the Rooch Framework.
 -  [`0x3::coin`](coin.md#0x3_coin)
 -  [`0x3::coin_store`](coin_store.md#0x3_coin_store)
 -  [`0x3::core_addresses`](core_addresses.md#0x3_core_addresses)
--  [`0x3::decoding`](decoding.md#0x3_decoding)
 -  [`0x3::ecdsa_k1`](ecdsa_k1.md#0x3_ecdsa_k1)
 -  [`0x3::ed25519`](ed25519.md#0x3_ed25519)
 -  [`0x3::empty`](empty.md#0x3_empty)
--  [`0x3::encoding`](encoding.md#0x3_encoding)
 -  [`0x3::ethereum_address`](ethereum_address.md#0x3_ethereum_address)
 -  [`0x3::ethereum_light_client`](ethereum_light_client.md#0x3_ethereum_light_client)
 -  [`0x3::ethereum_validator`](ethereum_validator.md#0x3_ethereum_validator)
 -  [`0x3::gas_coin`](gas_coin.md#0x3_gas_coin)
 -  [`0x3::genesis`](genesis.md#0x3_genesis)
--  [`0x3::hash`](hash.md#0x3_hash)
 -  [`0x3::multichain_address`](multichain_address.md#0x3_multichain_address)
 -  [`0x3::native_validator`](native_validator.md#0x3_native_validator)
 -  [`0x3::onchain_config`](onchain_config.md#0x3_onchain_config)
