+++ frameworks/rooch-framework/doc/ethereum_address.md
@@ -13,8 +13,8 @@
 -  [Function `into_bytes`](#0x3_ethereum_address_into_bytes)
 
 
-<pre><code><b>use</b> <a href="ecdsa_k1.md#0x3_ecdsa_k1">0x3::ecdsa_k1</a>;
-<b>use</b> <a href="hash.md#0x3_hash">0x3::hash</a>;
+<pre><code><b>use</b> <a href="">0x2::hash</a>;
+<b>use</b> <a href="ecdsa_k1.md#0x3_ecdsa_k1">0x3::ecdsa_k1</a>;
 </code></pre>
