+++ frameworks/rooch-framework/src/natives/rooch_framework/crypto/mod.rs
@@ -1,8 +1,5 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod decoding;
 pub mod ecdsa_k1;
 pub mod ed25519;
-pub mod encoding;
-pub mod hash;
